Linked between albums and photos, re-implement zip structure and now we have a short cut to load image data through server or local

+++ ITProject/ITProject/AppUtility/Util.swift
@@ -9,12 +9,14 @@
 import Foundation
 import FirebaseStorage
 import SVProgressHUD
+import SSZipArchive
 import Zip
+
 
 class Util {
     public static let BUTTON_DISMISS = "dismiss"
     
-    public static let EXTENSION_JPEG = ".jpg"
+    public static let EXTENSION_JPEG = "jpg"
     public static let EXTENSION_ZIP = ".zip"
     public static let IMAGE_FOLDER = "images"
     public static let TMP_FOLDER = "tmp"
@@ -119,15 +121,51 @@ class Util {
         }
     }
     
-    /* test image link   "795C8939-982E-40C8-AE2D-610A6EBA5866.jpg"
+    /*short cut to get Image from server by UID
+        imageUID : UID of the image
+        completion : do whatever you want to do when completion data : the image data
+        errorHandler : do whatever you want to do when error occur
+     */
+    public static func GetImageFromServer (imageUID: String,
+                                           completion: @escaping (Data?) -> () = { _ in },
+                                           errorHandler: @escaping (Error?) -> () = { _ in }){
+        let fullFliePath = Util.GetImagePathByUID(imageUID: imageUID)
+        
+        print("GetImageFromServer : did you get called twice ?")
+        DownloadFileFromServer(fileFullPath: fullFliePath, completion: { url in
+            
+            let fileLastPathUrl = (url?.lastPathComponent)! as NSString
+            let filenameWithNoExtension = fileLastPathUrl.deletingPathExtension
+            let fileExtension = "." + fileLastPathUrl.pathExtension
+            
+            print("Get Image from serever, Full url file : " + (fileLastPathUrl as String))
+            print("Get Image from serever , filenameWithNoExtension :" + filenameWithNoExtension)
+            print("Get Image from serever , fileExtension :" + fileExtension)
+            
+            Util.GetDataFromLocalFile(filename: filenameWithNoExtension, fextension: fileExtension, completion:
+                { data in
+                    completion(data)
+            })
+            
+        }, errorHandler: errorHandler)
+    }
+    
+    public static func GetImagePathByUID(imageUID:String) ->String{
+        return Util.IMAGE_FOLDER + "/" + imageUID + Util.EXTENSION_ZIP
+    }
+    
+    
+    /* test image link   "images/795C8939-982E-40C8-AE2D-610A6EBA5866.zip"
      error solver error 518:
      https://stackoverflow.com/questions/37470266/error-when-downloading-from-firebase-storage
      todo: get extension from database, here should be doing unzip
     */
     public static func DownloadFileFromServer (fileFullPath:String,
-                                               completion: @escaping (URL?) -> () = { _ in },
+                                               completion: @escaping (URL?) -> () = { _  in },
                                                errorHandler: @escaping (Error?) -> () = { _ in }){
         print("DownloadFileFromServer :file path " + fileFullPath)
+        
+        
         let gsReference = Storage.storage().reference(forURL: FIREBASE_STORAGE_URL + fileFullPath)
         
         let localURL = URL(fileURLWithPath: GetDocumentsDirectory().appendingPathComponent(fileFullPath).absoluteString)
@@ -147,16 +185,16 @@ class Util {
 //                print(Thread.current) // this is a UI thread
 
                 let fileAt = url!.deletingLastPathComponent().absoluteString
-                print("DownloadFileFromServer fileAt :" + fileAt)
+                print("DownloadFileFromServer local fileAt :" + fileAt)
                 let fileName = ((url!.lastPathComponent) as NSString)
-                print("DownloadFileFromServer fileName :" + (fileName as String))
+                print("DownloadFileFromServer  fileName :" + (fileName as String))
                 
                 UnzipFile(from: fileAt as NSString,
                           to: fileAt as NSString,
                           fileName: fileName as String,
-                          deleteAfterFinish: false){
-                            (unzippedFileURL) in
-                            completion(unzippedFileURL)
+                          deleteAfterFinish: true){
+                            filepath in
+                            completion(filepath)
                 }
             }
         }
@@ -232,34 +270,83 @@ class Util {
         }
     }
     
+    
+    public static func GetExtensionByFolderPath( fileFullPath: String)->String?{
+        if (fileFullPath.contains(Util.IMAGE_FOLDER + "/")){
+            return Util.EXTENSION_JPEG
+        }
+        return nil
+    }
+    
     /*
      Unzipfile function :
-     
          from : unzipfile location (folderPath)
          to : destination (folderPath)
          filename : the file need to be unzip, with extension .zip
          deleteAfterFinish : delete the zip file after finished unzipping
          completion handler : what you want to do after unzip, argument url : fullPath to the destination with filename. e.g.  file:/Users/xxxxxxxxxxxx/Documents/images/2E61DCE8-B133-4936-BDC7-E90FB4199B21.jpg */
     public static func UnzipFile(from: NSString, to: NSString,
-                                 fileName: String, deleteAfterFinish: Bool, completion: @escaping ((URL) -> Void) = {_ in }){
-        DispatchQueue(label: "working_queue", qos: .userInitiated).async {
-            let fullFilePath = from.appendingPathComponent(fileName)
-            
-            do {
-                try  Zip.unzipFile(URL(string: fullFilePath)!, destination: URL(string: to as String)!, overwrite: true, password: nil, progress: { (progress) -> () in
-                    print("UnzipFile : progress" + String(progress))
-                }){ unzippedFile in
-                    completion(unzippedFile)
+                                 fileName: String, deleteAfterFinish: Bool,
+                                completion: @escaping ((URL) -> Void) = {_ in }){
+
+        let fullFilePath = from.appendingPathComponent(fileName)
+        var unzippedFile = (fileName as NSString).deletingPathExtension
+        unzippedFile = (unzippedFile as NSString).appendingPathExtension(GetExtensionByFolderPath(fileFullPath: to as String)!)!
+
+        do {
+            try  Zip.unzipFile(URL(string: fullFilePath)!, destination: URL(string: to as String)!, overwrite: true, password: nil, progress: { (progress) -> () in
+                print("UnzipFile : progress " + String(progress))
+                if (progress >= 1.0){
+                    let unzipped = URL(string: (to.appendingPathComponent(unzippedFile)))
+                    completion(unzipped!)
+                    if deleteAfterFinish{
+                        do {
+                            try FileManager.default.removeItem(at: URL(string: fullFilePath)!)
+                        }catch let error as NSError{
+                            print ("UnzipFile : error occurs during remove unzip : " + error.localizedDescription)
+                        }
+                    }
                 }
-                
-                if deleteAfterFinish{
-                    try FileManager.default.removeItem(at: URL(string: fullFilePath)!)
-                }
-            }catch let error as NSError{
-                print ("UnzipFile : error occurs during unzip : " + error.localizedDescription)
-            }
+            })
+        }catch let error as NSError{
+            print ("UnzipFile : error occurs during unzip : " + error.localizedDescription)
         }
     }
+    
+    /* new unzip lib, the old one is too stupid, deprecated
+     from : unzipfile location (folderPath)
+     to : destination (folderPath)
+     filename : the file need to be unzip, with extension .zip
+     deleteAfterFinish : delete the zip file after finished unzipping
+     completion handler : what you want to do after unzip, argument url : fullPath to the destination with filename. e.g.  file:/Users/xxxxxxxxxxxx/Documents/images/2E61DCE8-B133-4936-BDC7-E90FB4199B21.jpg
+     */
+//    public static func UnzipFile(from: String, to: String,
+//                                 fileName: String, deleteAfterFinish: Bool,
+//                                 completion: @escaping ((String, Bool, Error?) -> Void) = {_, _, _ in }){
+//
+//        let fullFilePath = URL (string : from + fileName)?.path
+//        let toFullFilePath = URL (string : to)?.path
+//
+//        print("UnzipFile : preparing to unzip file : " + fullFilePath!)
+//        SSZipArchive.unzipFile(atPath: fullFilePath!, toDestination:toFullFilePath!, progressHandler: nil, completionHandler: {
+//            filePath, succeed, error in
+//
+//            if succeed {
+//                print("UnzipFile : success ~! ")
+//                completion(filePath, succeed, error)
+//
+//                if deleteAfterFinish{
+//                    do {
+//                        try FileManager.default.removeItem(at: URL(string: fullFilePath!)!)
+//                    }catch let error as NSError{
+//                        print ("UnzipFile : error occurs during remove unzip : " + error.localizedDescription)
+//                    }
+//                }
+//            }else{
+//                print("UnzipFile : unzip file fail: " + error.debugDescription)
+//            }
+//        })
+//    }
     
     /*note : read write file opearation, use worker thread to do it*/
     /*
@@ -322,6 +409,10 @@ class Util {
         completion Handler : do what you want with the file data with it, 1 args data
      */
     public static func GetDataFromLocalFile(filename: String, fextension: String, completion:@escaping (Data)->() = {_ in }){
+        
+        print ("GetDataFromLocalFile : filename : " + filename)
+        print ("GetDataFromLocalFile : fileExtension : " + fextension)
+        
         let folderPath = GetFolderByExtension(fextension: fextension, withPathSlash: true)!
         let fileDocumentFullPath = GetDocumentsDirectory().appendingPathComponent(folderPath, isDirectory: true).absoluteString as NSString
         
@@ -343,9 +434,10 @@ class Util {
             print("GetDataFromFile : going to unzip :" + fileDocumentFullPath.appendingPathComponent(zipFilename))
             
             if DoesFileExist(fullPath: fileDocumentFullPath.appendingPathComponent(zipFilename)){
-                UnzipFile(from: fileDocumentFullPath, to: fileDocumentFullPath, fileName: zipFilename, deleteAfterFinish: false){fileUrl in
-                    print("GetDataFromFile : unzip  with file url :" + fileUrl.absoluteString )
-                    ReadFileFromDocumentDirectory(fileName: GetFullFilePath(fileName: fileUrl.lastPathComponent)){
+                UnzipFile(from: fileDocumentFullPath, to: fileDocumentFullPath, fileName: zipFilename, deleteAfterFinish: false){url  in
+                    
+                    print("GetDataFromFile : unzip  with file url :" + url.absoluteString )
+                    ReadFileFromDocumentDirectory(fileName: GetFullFilePath(fileName: url.lastPathComponent)){
                         data in
                         print("GetDataFromFile : get file data success")
                         DispatchQueue.main.async {
@@ -379,7 +471,8 @@ class Util {
         withPathSlash: add pathSlash "/" at the end of the folder
      */
     public static func GetFolderByExtension(fextension: String, withPathSlash: Bool)->String?{
-        if (fextension) == (EXTENSION_JPEG) {
+        
+        if (fextension) == (EXTENSION_JPEG) || (fextension) == ("." + EXTENSION_JPEG) {
             return withPathSlash ? (IMAGE_FOLDER + "/") : IMAGE_FOLDER
         }
         return nil
